refactor: move pf login page components outside the 'pages' folder
